Logically separate broadcast layout and gallery/auto modes
