+++ frontend/src/aspects/Conference/Attend/Room/Vonage/Components/LayoutChooser.tsx
@@ -8,11 +8,7 @@ export default function LayoutChooser(): JSX.Element {
     const boxFill = useToken("colors", "gray.400");
     const insetBoxFillStroke = useToken("colors", "gray.600");
     const {
-        layout,
-        updateLayout,
-        saveLayout,
-        layoutChooser_isOpen: isOpen,
-        layoutChooser_onClose: onClose,
+        layout: { layout, updateLayout, saveLayout, layoutChooser_isOpen: isOpen, layoutChooser_onClose: onClose },
     } = useVonageLayout();
     const [initialLayout, setInitialLayout] = useState<{ layout: VonageSessionLayoutData; createdAt: number } | null>(
         null
