+++ frontend/src/aspects/Conference/Attend/Room/Vonage/VideoGrid/BroadcastLayouts/Single.tsx
@@ -1,0 +1,46 @@
+import { AspectRatio, Box, Center } from "@chakra-ui/react";
+import React from "react";
+import * as portals from "react-reverse-portal";
+import type { VizLayout_Single } from "../../Components/LayoutTypes";
+import StreamChooser from "../../Components/StreamChooser";
+
+export function Single({
+    visualLayout,
+    allowedToControlLayout,
+    isRecordingMode,
+}: {
+    visualLayout: VizLayout_Single;
+    allowedToControlLayout: boolean;
+    isRecordingMode: boolean;
+}): JSX.Element {
+    return (
+        <Center border="3px solid" borderColor="gray.900">
+            <AspectRatio
+                ratio={16 / 9}
+                w="100%"
+                maxW="calc((16 / 9) * (90vh - 250px))"
+                bgColor="gray.500"
+                mr={1}
+                pos="relative"
+            >
+                <Box h="100%" w="100%" bgColor="gray.500">
+                    {visualLayout.viewport ? (
+                        <portals.OutPortal
+                            node={visualLayout.viewport.component}
+                            enableVideo={true}
+                            resolution="high"
+                            framerate={30}
+                        />
+                    ) : undefined}
+                    {allowedToControlLayout ? (
+                        <StreamChooser
+                            positionKey="position1"
+                            centered={!visualLayout.viewport}
+                            isRecordingMode={isRecordingMode}
+                        />
+                    ) : undefined}
+                </Box>
+            </AspectRatio>
+        </Center>
+    );
+}
